+++ packages/xsd/lib/src/codecs/negativeInteger/negative_integer_codec.dart
