+++ packages/xsd/lib/src/codecs/unsignedShort/unsigned_short_codec.dart
